+++ src/lib/djerba/plugins/fusion/preprocess.py
@@ -230,6 +230,7 @@ class prepare_fusions(logger):
         """
         new_df = df[["Sample", "fusion_pairs"]].copy()
         new_df = self.drop_duplicates(new_df, ["fusion_pairs"])
+        new_df["fusion_pairs"] = new_df["fusion_pairs"].astype(str)
         new_df = new_df[~new_df["fusion_pairs"].str.contains("None")]
         # Change headers
         new_df.rename(columns={'Sample': 'Tumor_Sample_Barcode', 'fusion_pairs': 'Fusion'}, inplace=True)
@@ -256,17 +257,6 @@ class prepare_fusions(logger):
     
     def fix_reading_frames(self, df):
         """
-        Makes two columns: one with all reading frames (informative), and one that just says "Multiple Frames" if there are multiple.
-        Valid reading frames are in-frame, out-of-frame, and stop-codon.
-        According to the arriba documentation:
-        
-            reading_frame : This column states whether the gene at the 3' end of the fusion is fused in-frame or out-of-frame. 
-            The value stop-codon indicates that there is a stop codon prior to the fusion junction, such that the 3' end is not translated, 
-            even if the reading frame is preserved across the junction. The prediction of the reading frame builds on the prediction of the 
-            peptide sequence. A dot (.) indicates that the peptide sequence cannot be predicted, for example, because the transcript sequence 
-            could not be determined or because the breakpoint of the 5' gene does not overlap a coding region
-            
-        Anything else should be left as nan.
         """
         df["reading_frame"] = df["reading_frame"].replace(".", np.nan)
         return df
@@ -520,7 +510,17 @@ class prepare_fusions(logger):
 
         This function does not return anything.
         """
-        df_merged = self.merge_mavis_arriba(df_mavis, df_arriba)
+        
+        if df_mavis.empty:
+            # I.e. if all above filtering steps leave no fusions
+            # The usual culprit is that no fusions have greater than 20 reads
+
+            df_merged = self.left_join(df_mavis, df_arriba, "fusion_pairs")
+
+        else:
+            df_merged = self.merge_mavis_arriba(df_mavis, df_arriba)
+
+        #df_merged = self.merge_mavis_arriba(df_mavis, df_arriba)
 
         # Get the NCCN calls
         df_nccn = self.process_nccn(df_merged, oncotree_code)